backend for movie recommendation
