--- backend/requirements.txt
@@ -1,61 +1,0 @@
-altair==4.2.0
-anyio==3.6.1
-attrs==22.1.0
-autopep8==1.6.0
-blinker==1.5
-cachetools==5.2.0
-certifi==2022.9.24
-charset-normalizer==2.1.1
-click==8.1.3
-colorama==0.4.5
-commonmark==0.9.1
-decorator==5.1.1
-entrypoints==0.4
-fastapi==0.79.0
-Flask==2.2.2
-gitdb==4.0.9
-GitPython==3.1.28
-h11==0.13.0
-idna==3.3
-importlib-metadata==5.0.0
-itsdangerous==2.1.2
-Jinja2==3.1.2
-jsonschema==4.16.0
-lxml==4.9.1
-MarkupSafe==2.1.1
-numpy==1.23.1
-packaging==21.3
-pandas==1.4.3
-Pillow==9.2.0
-protobuf==3.20.3
-pyarrow==9.0.0
-pycodestyle==2.9.1
-pydantic==1.9.1
-pydeck==0.8.0b3
-Pygments==2.13.0
-Pympler==1.0.1
-pyparsing==3.0.9
-pyrsistent==0.18.1
-python-dateutil==2.8.2
-pytz==2022.1
-pytz-deprecation-shim==0.1.0.post0
-requests==2.28.1
-rich==12.6.0
-semver==2.13.0
-six==1.16.0
-smmap==5.0.0
-sniffio==1.2.0
-starlette==0.19.1
-streamlit==1.13.0
-toml==0.10.2
-toolz==0.12.0
-tornado==6.2
-typing_extensions==4.3.0
-tzdata==2022.4
-tzlocal==4.2
-urllib3==1.26.12
-uvicorn==0.18.2
-validators==0.20.0
-watchdog==2.1.9
-Werkzeug==2.2.2
-zipp==3.9.0
